+++ packages/angular/CHANGELOG.md
@@ -1,3 +1,6 @@
+# [Unreleased]
+- Fix RequestHeadersInterceptor not to overwrite `Authorization` request header if it was already set
+
 # 0.0.2-alpha.4 (2021-06-10)
 - Enable JWT in VcdApiClient
 - Export vcd.transfer.client through the client index